fold cpu hadamard into activation transform
