--- crates/backend-uzu/src/backends/cpu/kernel/hadamard_transform/hadamard_transform.rs
@@ -1,20 +1,0 @@
-use crate::backends::common::gpu_types::HADAMARD_TRANSFORM_BLOCK_SIZE;
-
-pub(crate) fn hadamard_transform(values: &mut [f32; HADAMARD_TRANSFORM_BLOCK_SIZE]) {
-    let mut stride = 1;
-    while stride < HADAMARD_TRANSFORM_BLOCK_SIZE {
-        for lane in 0..HADAMARD_TRANSFORM_BLOCK_SIZE {
-            if lane & stride == 0 {
-                let a = values[lane];
-                let b = values[lane | stride];
-                values[lane] = a + b;
-                values[lane | stride] = a - b;
-            }
-        }
-        stride <<= 1;
-    }
-    let scale = 1.0 / (HADAMARD_TRANSFORM_BLOCK_SIZE as f32).sqrt();
-    for v in values.iter_mut() {
-        *v *= scale;
-    }
-}
