--- crates/backend-uzu/src/backends/cpu/kernel/hadamard_transform/mod.rs
@@ -1,1 +1,0 @@
-pub mod hadamard_transform;
