+++ crates/backend-uzu/src/backends/cpu/kernel/mod.rs
@@ -8,7 +8,6 @@ mod attention;
 mod embedding;
 mod gated_act_mul;
 mod gdn;
-mod hadamard_transform;
 mod logit_soft_cap;
 mod matmul;
 mod moe;
